Working on shop navigation styling
