--- src/components/ShopDropdownMenu/Styles.ts
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-export const Img = styled.img`
-  border-radius: 50%;
-  height: 36px;
-  width: 36px;
-
-  &:hover {
-    cursor: pointer;
-  }
-`;
